+++ frontend/lib/utils/rule.ts
@@ -22,6 +22,27 @@ export function normalizeRuleActions(
   }));
 }
 
+// Helper to determine effectiveScope and effectiveFromDate from effective_from string
+export function getEffectiveScopeAndDate(effective_from?: string): {
+  effectiveScope: "all" | "from";
+  effectiveFromDate: Date | undefined;
+} {
+  if (effective_from) {
+    const effDate = new Date(effective_from);
+    // Check for invalid date
+    if (isNaN(effDate.getTime())) {
+      return { effectiveScope: "all", effectiveFromDate: undefined };
+    }
+    const isToday = effDate.toDateString() === new Date().toDateString();
+    if (isToday) {
+      return { effectiveScope: "all", effectiveFromDate: undefined };
+    } else {
+      return { effectiveScope: "from", effectiveFromDate: effDate };
+    }
+  }
+  return { effectiveScope: "all", effectiveFromDate: undefined };
+}
+
 export function normalizeRuleConditions(
   conditions: BaseRuleCondition[],
   categories: Category[]
